#48, Update user droppoint with admin id bug fixed

+++ TaskCat/TaskCat/Controller/DropPointController.cs
@@ -16,6 +16,9 @@
     using System.Threading.Tasks;
     using System.Web.Http;
 
+    /// <summary>
+    /// Registers and manages drop points for a certain user
+    /// </summary>
     public class DropPointController : ApiController
     {
         private IDropPointService service;
@@ -36,20 +39,22 @@
         [Authorize]
         public async Task<IHttpActionResult> Get(string query, string userId = null, int pageSize = AppConstants.DefaultPageSize, int page = 0, bool envelope = true)
         {
-            if(string.IsNullOrWhiteSpace(query))
+            if (string.IsNullOrWhiteSpace(query))
             {
                 throw new ArgumentNullException(nameof(query));
             }
 
             var currentUserId = this.User.Identity.GetUserId();
 
-            if (userId!=null && currentUserId != userId 
+            if (userId != null && currentUserId != userId
                 && (!this.User.IsInRole(RoleNames.ROLE_ADMINISTRATOR) || !this.User.IsInRole(RoleNames.ROLE_BACKOFFICEADMIN)))
             {
                 // TODO: Need to fix this differently by a proper result
                 return Unauthorized();
             }
-            userId = currentUserId;
+
+            if (string.IsNullOrWhiteSpace(userId))
+                userId = currentUserId;
 
             pageSize = pageSize > AppConstants.MaxPageSize ? AppConstants.MaxPageSize : pageSize;
             var queryResult = await service.SearchDropPoints(userId, query);
@@ -107,34 +112,67 @@
 
 
             if (value.UserId != null && value.UserId != authorizedId
-                && (!this.User.IsInRole("Administrator") || !this.User.IsInRole("BackOfficeAdmin")))
+                && (!this.User.IsInRole(RoleNames.ROLE_ADMINISTRATOR) || !this.User.IsInRole(RoleNames.ROLE_BACKOFFICEADMIN)))
             {
                 // TODO: Need to fix this differently by a proper result
                 return Unauthorized();
             }
+
             value.Id = default(string);
-            value.UserId = authorizedId;
+            if (string.IsNullOrWhiteSpace(value.UserId))
+                value.UserId = authorizedId;
             var result = await service.Insert(value);
             return Json(result);
         }
 
         [HttpPut]
+        [Authorize]
         public async Task<IHttpActionResult> Put(DropPoint value)
         {
             if (!ModelState.IsValid)
             {
                 return BadRequest(ModelState);
             }
-            var result = await service.Update(value);
-            return Json(result);
+
+            var authorizedId = User.Identity.GetUserId();
+
+            if (value.UserId != null && value.UserId != authorizedId
+                && (!this.User.IsInRole(RoleNames.ROLE_ADMINISTRATOR) || !this.User.IsInRole(RoleNames.ROLE_BACKOFFICEADMIN)))
+            {
+                // TODO: Need to fix this differently by a proper result
+                return Unauthorized();
+            }
+
+            if (string.IsNullOrWhiteSpace(value.UserId))
+                value.UserId = authorizedId;
+         
+            if(this.User.IsInRole(RoleNames.ROLE_ADMINISTRATOR) || this.User.IsInRole(RoleNames.ROLE_BACKOFFICEADMIN))
+            {
+                return Json(await service.Update(value));
+            }
+            else
+            {
+                return Json(await service.Update(value, value.UserId));
+            }
+            
         }
 
         [HttpDelete]
         [Authorize]
-        public async Task<DropPoint> Delete(string id)
+        public async Task<IHttpActionResult> Delete(string id, string userId = null)
         {
+            var authorizedId = User.Identity.GetUserId();
+            if (userId != null && userId != authorizedId
+                && (!this.User.IsInRole(RoleNames.ROLE_ADMINISTRATOR) || !this.User.IsInRole(RoleNames.ROLE_BACKOFFICEADMIN)))
+            {
+                // TODO: Need to fix this differently by a proper result
+                return Unauthorized();
+            }
+
+            if (string.IsNullOrWhiteSpace(userId))
+                userId = authorizedId;
             var result = await service.Delete(id);
-            return result;
+            return Json(result);
         }
     }
 }
